requirements.txt
firebase-admin
speedtest-cli
